change order of tabs
